Updates to exercise 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,15 +9,15 @@
 int main(void)
 {
 	int i;
-	double num1 = 1;
-	double num2 = 2;
-	double sum;
+	unsigned int num1 = 1;
+	unsigned int num2 = 2;
+	unsigned int sum;
 
-	printf("%lf, %lf, ", num1, num2);
+	printf("%lu, %lu, ", num1, num2);
 	for (i = 0; i < 96; i++)
 	{
 		sum = num1 + num2;
-		printf("%lf", sum);
+		printf("%lu", sum);
 		if (i != 95)
 			printf(", ");
 		num1 = num2;
